refactor: extract middleware factory from api/app.py
